KFSPTS-7312: More changes, but still not fully working.
