+++ src/main/java/edu/cornell/kfs/concur/rest/xmlObjects/AccessTokenDTO.java
@@ -15,6 +15,8 @@ public class AccessTokenDTO {
     protected String token;
     @XmlElement(name = "Expiration_Date")
     protected String expirationDate;
+    @XmlElement(name = "Refresh_Token")
+    protected String refreshToken;
     
     public String getInstanceURL() {
         return instanceURL;
@@ -33,6 +35,12 @@ public class AccessTokenDTO {
     }
     public void setExpirationDate(String expirationDate) {
         this.expirationDate = expirationDate;
+    }
+    public String getRefreshToken() {
+        return refreshToken;
+    }
+    public void setRefreshToken(String refreshToken) {
+        this.refreshToken = refreshToken;
     }
 
 }
